Change default on disableHostDeletion to true
Currently the example yaml disables host deletion which causes
hosts to disappear from a Kubernetes cluster if the host does not
have volumes attached. Administrators then need to manually
re-add the host to make it available to the cluster again.
This commit changes this behaviour to a more user-friendly default.